Use `dev-dependencies`
