+++ rumqttc/src/broker.rs
@@ -1,5 +1,5 @@
-use rumqttc::mqttbytes::v4::*;
-use rumqttc::mqttbytes::*;
+use crate::mqttbytes::v4::*;
+use crate::mqttbytes::*;
 use std::collections::VecDeque;
 use std::io;
 use std::time::Duration;
@@ -7,11 +7,12 @@ use tokio::net::TcpListener;
 use tokio::select;
 use tokio::{task, time};
 
+use crate::{Event, Incoming, Outgoing, Packet};
 use bytes::BytesMut;
 use flume::{bounded, Receiver, Sender};
-use rumqttc::{Event, Incoming, Outgoing, Packet};
 use tokio::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
 
+/// Dummy broker used in unit tests
 pub struct Broker {
     pub(crate) framed: Network,
     pub(crate) incoming: VecDeque<Packet>,
@@ -42,7 +43,7 @@ impl Broker {
                             incoming,
                             outgoing_tx,
                             outgoing_rx,
-                        }
+                        };
                     }
                 };
 
@@ -58,6 +59,35 @@ impl Broker {
             incoming: VecDeque::new(),
             outgoing_tx,
             outgoing_rx,
+        }
+    }
+
+    // Reads a publish packet from the stream with 2 second timeout
+    pub async fn read_publish(&mut self) -> Option<Publish> {
+        loop {
+            let packet = if !self.incoming.is_empty() {
+                self.incoming.pop_front().unwrap()
+            } else {
+                let packet = time::timeout(Duration::from_secs(2), async {
+                    self.framed.readb(&mut self.incoming).await.unwrap();
+                    self.incoming.pop_front().unwrap()
+                })
+                .await;
+
+                match packet {
+                    Ok(packet) => packet,
+                    Err(_e) => return None,
+                }
+            };
+
+            match packet {
+                Packet::Publish(publish) => return Some(publish),
+                Packet::PingReq => {
+                    self.framed.write(Packet::PingResp).await.unwrap();
+                    continue;
+                }
+                packet => panic!("Expecting a publish. Received = {:?}", packet),
+            }
         }
     }
 
@@ -79,6 +109,12 @@ impl Broker {
         loop {
             self.framed.readb(&mut self.incoming).await.unwrap();
         }
+    }
+
+    /// Sends an acknowledgement
+    pub async fn ack(&mut self, pkid: u16) {
+        let packet = Packet::PubAck(PubAck::new(pkid));
+        self.framed.write(packet).await.unwrap();
     }
 
     /// Sends an acknowledgement
